Moves COPY.json to client
