+++ client/app/queue/AttorneyTaskListView.jsx
@@ -18,7 +18,7 @@ import {
 import { clearCaseSelectSearch } from '../reader/CaseSelect/CaseSelectActions';
 
 import { fullWidth } from './constants';
-import COPY from '../../../COPY.json';
+import COPY from '../../COPY.json';
 
 class AttorneyTaskListView extends React.PureComponent {
   componentWillUnmount = () => {
