+++ client/app/queue/JudgeReviewTaskListView.jsx
@@ -19,7 +19,7 @@ import {
 import { clearCaseSelectSearch } from '../reader/CaseSelect/CaseSelectActions';
 
 import { fullWidth } from './constants';
-import COPY from '../../../COPY.json';
+import COPY from '../../COPY.json';
 
 class JudgeReviewTaskListView extends React.PureComponent {
   componentWillUnmount = () => {
